% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter.R
\name{butter_classify}
\alias{butter_classify}
\title{classify}
\usage{
butter_classify(washedDat, classList)
}
\arguments{
\item{washedDat}{result of wash()}

\item{classList}{result of pipeButter()}
}
\value{
matrix of classifier results
}
\description{
classify, wrapper to sc_classify
}
