% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnstatus.R
\name{cn_barplot_grnSing_base}
\alias{cn_barplot_grnSing_base}
\title{barplot this specific GRN}
\usage{
cn_barplot_grnSing_base(
  qScores,
  ctrlScores,
  stQuery,
  dLevelQ,
  snName,
  ctrSamps,
  bOrder,
  sidCol = "sample_id"
)
}
\arguments{
\item{qScores}{queryScores}

\item{stQuery}{sample table}

\item{dLevelQ}{dLevel of query samples}

\item{snName}{name of subnet to plot establishment level}

\item{ctrSamps}{names of samples in training data}

\item{bOrder}{order of bars}

\item{sidCol}{sample id colname}

\item{control}{scores}
}
\value{
gpplot barplot
}
\description{
barplot this specific GRN
}
