% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_assess.R
\name{cn_computeAUCPR}
\alias{cn_computeAUCPR}
\title{compute AUPCR}
\usage{
cn_computeAUCPR(perfDF, precisionCol = "Precision",
  recallCol = "Recall", predCol = "Predictions")
}
\arguments{
\item{perfDF}{a df}

\item{precisionCol}{"Precision"}

\item{recallCol}{"Recall"}

\item{predCol}{"Predictions"}
}
\value{
auprs
}
\description{
compute AUPCR
}
