% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cn_correctZmat}
\alias{cn_correctZmat}
\title{make Inf and -Inf values sensible}
\usage{
cn_correctZmat(zmat)
}
\arguments{
\item{zMat}{zMat}
}
\value{
corrected zMat
}
\description{
make Inf and -Inf values sensible
}
