% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_assess.R
\name{cn_eval}
\alias{cn_eval}
\title{run cn_clPerf across thresholds}
\usage{
cn_eval(vect, sampTab, dLevel, classification, threshs = seq(0, 1, by =
  0.05), dLevelSID = "sample_id")
}
\arguments{
\item{vect}{named vector}

\item{dLevel}{description level}

\item{classification}{classification matrix}

\item{threshs}{seq of pval cutoffs}

\item{dLevelSID}{column to indicate sample id}
}
\value{
return a data frame of the number of TP, FN, FP, and TN, and pval cutoff
}
\description{
run cn_clPerf across thresholds
}
