% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_assess.R
\name{cn_findSensAt}
\alias{cn_findSensAt}
\title{return the sens at given FPR}
\usage{
cn_findSensAt(rocRes, at = 0.05)
}
\arguments{
\item{rocRes}{result of running cn_eval}

\item{what}{is sens at this FPR?}
}
\value{
what is sens at selected FPR?
}
\description{
return the sens at given FPR
}
