% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_assess.R
\name{cn_rectArea}
\alias{cn_rectArea}
\title{compute area of rect given by}
\usage{
cn_rectArea(ptA, ptB, ptC)
}
\arguments{
\item{ptA}{point a}

\item{ptB}{point b}

\item{ptC}{point C}
}
\value{
area
}
\description{
compute area of rect given by
}
