% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{compareGenePairs}
\alias{compareGenePairs}
\title{Compile genePairs comparison matrix}
\usage{
compareGenePairs(query_exp, training_exp, training_st, classCol,
  sampleCol = NULL, RF_classifier, numPairs = 20,
  trainingOnly = TRUE)
}
\arguments{
\item{query_exp}{the expression matrix of query samples}

\item{training_exp}{the expression matrix of training samples}

\item{training_st}{the sample table of training data}

\item{classCol}{the column name of the sample table that indicates the classes}

\item{sampleCol}{the column name of the sample table that indicates the sample names. NULL if sample names are indcated in rownames of the sample table}

\item{numPairs}{the number of genes to extract for comparison from most important to least important in the classifier}

\item{cnProc}{the cnProc from the training}
}
\value{
gene pair comparison matrix
}
\description{
compile genePairs comparison matrix for users to elucidate the biological significance behind classification results
}
