% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{compileGSEA}
\alias{compileGSEA}
\title{compileGSEA}
\usage{
compileGSEA(gseaEnr, thresh = 0.05)
}
\arguments{
\item{gseaEnr}{from rnunning gsea.wrapper.set}

\item{thresh}{defaults 0.05}
}
\value{
df df
}
\description{
make a long table from fgsea.wrapper.set
}
\details{
make a long table from fgsea.wrapper.set
}
