% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{csRenameOrth}
\alias{csRenameOrth}
\title{enable cross-species comparison}
\usage{
csRenameOrth(expQuery, expTrain, orthTable, speciesQuery = "human",
  speciesTrain = "mouse")
}
\arguments{
\item{expQuery}{expQuery}

\item{expTrain}{expTrain}

\item{orthTable}{orthTable}

\item{speciesQuery}{speciesQuery default human}

\item{speciesTrain}{speciesTrain default mouse}
}
\value{
list (expQuery, expTrain)
}
\description{
rename and subset query expDat to orthologs in training data
}
