% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter.R
\name{divide_sampTab}
\alias{divide_sampTab}
\title{divide sample table}
\usage{
divide_sampTab(sampTab, prop = 0.5, dLevel = "prefix")
}
\arguments{
\item{sampTab}{sampTab}

\item{prop}{prop}

\item{dLevel}{dLevel}
}
\value{
list of stTrain stVal
}
\description{
divide sample table
}
