% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steam.R
\name{.steam_dt}
\alias{.steam_dt}
\title{butt_dt dynamic tree cluster}
\usage{
.steam_dt(sampTab, hcRes, cors, minModSize = 20, deepSplit = FALSE)
}
\arguments{
\item{sampTab}{sampTab}

\item{minModSize}{minimum module size}

\item{deepSplit}{deepSplit}

\item{datMat}{data matrix to cluster, cols are samples/cells}

\item{distanceMetric}{distanceMetric}

\item{linkageMetric}{linkageMetric}

\item{maxTreeHeight}{maxTreeHeight depends on}
}
\value{
sampTab with group column indicating cluster assignment
}
\description{
butt_dt dynamic tree cluster
}
