% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downSampleW}
\alias{downSampleW}
\title{weighted subtraction from mapped reades}
\usage{
downSampleW(vector, total = 1e+05, dThresh = 0)
}
\arguments{
\item{vector}{of total mapped reads per gene/transcript}

\item{total}{post transformation sum of read counts}
}
\value{
vector of downsampled read mapped to genes/transcripts
}
\description{
Simulate expression profile of  _total_ mapped reads
}
