% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractSCE}
\alias{extractSCE}
\title{extract sampTab and expDat sce object into regular S3 objects}
\usage{
extractSCE(sce_object, exp_slot_name = "counts")
}
\description{
extract sampTab and expDat sce object into regular S3 objects
}
