% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{fgsea.wrapper.set}
\alias{fgsea.wrapper.set}
\title{run fgsea on list of diffExprs}
\usage{
fgsea.wrapper.set(diffExprList, geneSets, minSize = 10, nPerm = 10000)
}
\arguments{
\item{diffExprList}{named list of diffExpr results, for example from running gnrAll()}

\item{geneSets}{named list of gene sets}

\item{minSize}{defaults to 10}

\item{nPerm}{defaults to 1e-4}
}
\value{
list of and list of ks.wrapper results
}
\description{
find gene sets enriched in provided ranked gene list
}
