% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_genes_byGo}
\alias{find_genes_byGo}
\title{find transcript factors}
\usage{
find_genes_byGo(annotation, species = "Hs", onto = "BP")
}
\arguments{
\item{species}{defaul is 'Hs', can also be 'Mm;}

\item{ontology}{default is BP}
}
\value{
vector fo TF names
}
\description{
find transcript factors
}
