% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{find_top_pc}
\alias{find_top_pc}
\title{find topX PCs}
\usage{
find_top_pc(vp, qtile = 0.25)
}
\arguments{
\item{vp}{result of running vg_pca}

\item{qtile}{quantile}
}
\value{
index of PC at which _quantile_ of sd is included
}
\description{
find topX PCs .don't use this
}
