% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{findAvgLabel}
\alias{findAvgLabel}
\title{Make training label}
\usage{
findAvgLabel(gpTab, stTrain, dLevel)
}
\arguments{
\item{gpTab}{}

\item{stTrain}{sample table for training data}

\item{dLevel}{the annotation for cell type label}
}
\value{
a named vector for plotting purposes
}
\description{
To make name vector for plotting purpose
}
