% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{findBestPairs}
\alias{findBestPairs}
\title{finds the best pairs to use}
\usage{
findBestPairs(xdiff, n = 50, maxPer = 3)
}
\arguments{
\item{xdiff}{xdiff}

\item{n}{number of pairs}

\item{maxPer}{indicates the number of pairs that a gene is allowed to be in}
}
\value{
vector of good pairs
}
\description{
finds the best pairs to use
}
