% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GEP_makeMean}
\alias{GEP_makeMean}
\title{row average (or median) based on groups}
\usage{
GEP_makeMean(exp, groupings, type = "mean")
}
\arguments{
\item{exp}{expression df}

\item{groupings}{groupings}

\item{type}{mean or media}
}
\value{
return a dataframe of mean or median-ed data based on given groupings.  colnames become the column name of the first sample in each group from the original data
}
\description{
row average (or median) based on groups
}
