% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{getClassGenes}
\alias{getClassGenes}
\title{Find Classy Genes}
\usage{
getClassGenes(diffRes, topX = 25, bottom = TRUE)
}
\arguments{
\item{diffRes}{a dataframe with pval, cval, holm, and rownames as the gene names}

\item{topX}{a number dicataing the number of genes to select for training classifier}

\item{bottom}{logic if true use the top x genes with - cvals}
}
\value{
a vector of genes that are good for training classifier for that category
}
\description{
Extract genes suitable for training classifier
}
