% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getSpecGenes}
\alias{getSpecGenes}
\title{get the genes specific to each diffexp in a list of them}
\usage{
getSpecGenes(xdatList, topN = 50)
}
\arguments{
\item{xdatList}{list of diffexp data frame, must have cval column}

\item{topN}{50 numb of genes to return per diffexp}
}
\value{
named list of genes
}
\description{
get the genes specific to each diffexp in a list of them
}
