% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{gnrAll}
\alias{gnrAll}
\title{find genes higher in a cluster compared to all other cells}
\usage{
gnrAll(expDat, cellLabels)
}
\arguments{
\item{expDat}{expDat}

\item{cellLabels}{named vector of cell groups}
}
\value{
list of diffExp data framnes
}
\description{
ind genes higher in a cluster compared to all other cells
}
