% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{gnrBP}
\alias{gnrBP}
\title{find best pairs}
\usage{
gnrBP(expDat, cellLabels, topX = 50)
}
\arguments{
\item{expDat}{expDat}

\item{cellLabels}{named vector of cell groups}
}
\value{
vector of pairs
}
\description{
find best pairs
}
