% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpa.R
\name{gpa_recurse}
\alias{gpa_recurse}
\title{find clusters, agnostically (kind of)}
\usage{
gpa_recurse(expAll, kvals = 2:5, nPCs = 2, dThresh = 0, zThresh = 2,
  meanType = "overall_mean", maxLevel = 4, max.iter = 30,
  SilhDrop = 0.25, minClusterSize = 42, methods = c("mclust", "cutree",
  "kmeans", "sNN_clust"), pcaMethod = "prcomp", numGenes = 5,
  silMin = TRUE, pcAuto = TRUE, adjust = TRUE)
}
\arguments{
\item{expAll}{expDat}

\item{kvals}{kvals 2:5}

\item{nPCs}{number of PCs to use (only used if pcAuto=FALSE)}

\item{dThresh}{0,}

\item{zThresh}{2,}

\item{meanType}{="overall_mean",}

\item{maxLevel}{maxLevel =4,}

\item{max.iter}{30,}

\item{SilhDrop}{0.25,}

\item{minClusterSize}{=42,}

\item{methods}{methods =c("mclust", "cutree", "kmeans"),}

\item{pcaMethod}{"prcomp",}

\item{numGenes}{(5)}

\item{silMin}{terminate clustering on a set of cells if the current min silh is < threshold. in contrast to using mean silh}

\item{pcAuto}{TRUE auto select the best PCs based on difference in Stdev}

\item{adjust}{adjust TRUE whether to adjust silhouette based on the cluster size}
}
\value{
a very complex object
}
\description{
find clusters, agnostically (kind of). Run PAM and cutree, across range of ks, select best based on silhouette
}
