% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{harmonize}
\alias{harmonize}
\title{harmonize data sets}
\usage{
harmonize(expList, trans, dsVal = 1e+06, pres = TRUE)
}
\arguments{
\item{expList}{list of expression matricies}

\item{dsVal}{target expression count}

\item{pres}{bool whehter to report prob of expression highest 2 classes; default or not only highest class}
}
\value{
list1=list of normalized expression matricies, list2=list of prob of exp matricies
}
\description{
harmonize data sets
}
