% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hm_enr}
\alias{hm_enr}
\title{heatmap of the enrichment result}
\usage{
hm_enr(esList, threshold = 0.05, cRows = FALSE, cCols = FALSE, fsr = 4)
}
\arguments{
\item{esList}{returned from ks.extract.more}

\item{threshold}{threshold 0.05 (holm-corrected pval)}
}
\value{
nothing
}
\description{
Heatmap of tthe enrichment result
}
\examples{
hm_enr(esList)

}
