% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{hmClass}
\alias{hmClass}
\title{heatmap of the classification result}
\usage{
hmClass(classMat, isBig = FALSE, cluster_cols = FALSE)
}
\arguments{
\item{classMat}{classMat}

\item{isBig}{is this a big heatmap? TRUE or FALSE}

\item{cluster_cols}{cluster_cols}
}
\value{
nothing
}
\description{
Heatmap of the classification result.
}
\examples{
cn_HmClass(cnRes, isBig=TRUE)

}
