% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{ks.extract.more}
\alias{ks.extract.more}
\title{extract a matrix of ES scores and -1log10(pvals) from rthe result of ks.wrapper.set}
\usage{
ks.extract.more(enrRes, sigType = "holm", sigThresh = 1e-05,
  gsColName = "geneSet", esColName = "ES")
}
\arguments{
\item{enrRes}{for example from running ks.wrapper.set()}

\item{sigType}{defaults to Holm}

\item{sigThresh}{defaults to 1e-5}
}
\value{
atrix of ES for significant geneSets (cols=queryvectrs, rows=genesets),
}
\description{
get a matrix of ES scores for easy display
}
