% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{ks.wrapper}
\alias{ks.wrapper}
\title{run KS test given a ranked vector of genes and list of gene sets}
\usage{
ks.wrapper(queryGenes, geneSets)
}
\arguments{
\item{queryGenes}{ranked gene list as a vector}

\item{geneSets}{named list of gene sets}
}
\value{
data frame of geneSet name, enrichment score, stat, nominal pvalue, holm-corrected pvalue, fdr, ?comma separated character of leading edge genes?
}
\description{
find gene sets enriched in provided ranked gene list
}
