% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{ks.wrapper.set}
\alias{ks.wrapper.set}
\title{run ks.wrapper on list of diffExprs}
\usage{
ks.wrapper.set(diffExprList, geneSets)
}
\arguments{
\item{diffExprList}{named list of diffExpr results, for example from running gnrAll()}

\item{geneSets}{named list of gene sets}

\item{sigType}{defaults to Holm}

\item{sigThresh}{defaults to 1e-5}
}
\value{
list of and list of ks.wrapper results
}
\description{
find gene sets enriched in provided ranked gene list
}
