% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter.R
\name{makeClassifier}
\alias{makeClassifier}
\title{make a single RF classifier}
\usage{
makeClassifier(trainScores, ctt, stTrain, nTrees = 200,
  dLevel = "description1")
}
\arguments{
\item{trainScores}{numeric matrix of predictors}

\item{ctt}{what cell type/tissues is this?}

\item{stTrain}{sample table}
}
\value{
randomForest classifier
}
\description{
uses R's randomForest package
}
