% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter.R
\name{makeRFs}
\alias{makeRFs}
\title{make classifiers}
\usage{
makeRFs(expTrain, stTrain, geneLists, dLevel = "description1", nTrees = 200)
}
\arguments{
\item{expTrain}{expression matrix}

\item{stTrain}{sample table}

\item{dLevel}{stTrain column name to split on}

\item{list}{of dLevel=>vector of genes}
}
\value{
list of classifiers
}
\description{
Make one Random Forest classification per unique dLevel
}
