% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_train.R
\name{mclust_Mat}
\alias{mclust_Mat}
\title{run Mclust on an expression matrix}
\usage{
mclust_Mat(expDat, pres = TRUE)
}
\arguments{
\item{expDat}{matrix}

\item{pres}{boolean}
}
\value{
matrix rows=genes, cols=cells values
}
\description{
run Mclust on an expression matrix
}
