% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeLoad10x}
\alias{mergeLoad10x}
\title{Loads 10x data mtx}
\usage{
mergeLoad10x(
  basePath,
  pNames,
  nCells = 1000,
  secPath = "filtered_matrices_mex/hg19"
)
}
\arguments{
\item{basePath}{path}

\item{pNames}{are prefix labels to facilitate merging data sets when UMIs might collide}

\item{nCells}{number of cells per mtx to load}

\item{secPath}{secPath}
}
\value{
list of list(sampTab=sampTab, expDat=expDat)
}
\description{
Loads 10x data mtx
}
