% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{pair_transform}
\alias{pair_transform}
\title{makes complete gene-to-gene comparison}
\usage{
pair_transform(expDat)
}
\arguments{
\item{expDat}{expDat}
}
\value{
list of list(genes=data frame, expDat=binary matrix)
}
\description{
makes complete gene-to-gene comparison
}
