% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pca_to_tsne}
\alias{pca_to_tsne}
\title{make tsne from pca}
\usage{
pca_to_tsne(expDat, recRes, perplexity = 30, theta = 0.3, weighted = TRUE)
}
\arguments{
\item{expDat}{expDat}

\item{recRes}{result of running gpaRecurse}

\item{perplexity}{(30)}

\item{theta}{(0.30)}

\item{weighted}{(TRUE) whether to use PCs from deeper lelves, and to weight them}
}
\value{
tsne matrix
}
\description{
make tsne from pca
}
