% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_assess.R
\name{plot_class_PRs}
\alias{plot_class_PRs}
\title{Plot results of sc_classAssess}
\usage{
plot_class_PRs(assessed)
}
\arguments{
\item{assessed}{result of runnung cn_classAssess}
}
\value{
ggplot pbject
}
\description{
Plot one precision recall curve per CT
}
\examples{
testAssTues<-cn_splitMakeAssess(stTrain, expTrain, ctGRNs, prop=.5)
plot_class_PRs(testAssTues$ROCs)

}
