% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca_gpa}
\alias{plot_pca_gpa}
\title{plot gpa res}
\usage{
plot_pca_gpa(xtree, legend = FALSE)
}
\arguments{
\item{xtree}{xtree}

\item{legend}{whether to display it}
}
\value{
ggplot
}
\description{
plot gpa res
}
