% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDBscan}
\alias{plotDBscan}
\title{plot tsne results}
\usage{
plotDBscan(mcRes, legend = FALSE)
}
\arguments{
\item{mcRes, }{which must have TSNE.1 and TSNE.2, and group columns}

\item{legend}{whether to display it}
}
\value{
ggplot
}
\description{
plot tsne results
}
