% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGeneComparison}
\alias{plotGeneComparison}
\title{Gene expression plotting}
\usage{
plotGeneComparison(
  expDat,
  grps,
  fontsize_row = 6,
  cRows = FALSE,
  cCols = FALSE,
  toScale = FALSE,
  ...
)
}
\arguments{
\item{expDat}{comparison expression matrix from \code{\link{makeGeneCompareTab}}}

\item{fontsize_row}{row font size}

\item{cRows}{TRUE if cluster the rows}

\item{cCols}{TRUE if cluster the columns}
}
\value{
a heatmap of gene expressions for the comparsion expression matrix
}
\description{
Plot gene expressions for visualization and comparison
}
