% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGPALevel}
\alias{plotGPALevel}
\title{plot gpa res}
\usage{
plotGPALevel(gpaRes, llevel = "L1_G1", legend = FALSE)
}
\arguments{
\item{gpaRes}{gpRes}

\item{llevel}{llevel}

\item{legend}{whether to display it}
}
\value{
ggplot
}
\description{
plot gpa res
}
