% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGPApca}
\alias{plotGPApca}
\title{plot gpa res}
\usage{
plotGPApca(xres, legend = FALSE)
}
\arguments{
\item{legend}{whether to display it}

\item{gpaRes}{gpRes}
}
\value{
ggplot
}
\description{
plot gpa res
}
