% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGPArecTrans}
\alias{plotGPArecTrans}
\title{plot gpa_recurse res transform coordinates}
\usage{
plotGPArecTrans(gpaRes, pcaLevel = 1, grpLevel = 1, legend = FALSE)
}
\arguments{
\item{legend}{whether to display it}
}
\value{
ggplot
}
\description{
plot gpa_recurse res
}
