% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wash.R
\name{prewash}
\alias{prewash}
\title{down samples, computes per gene stats, finds ok genes}
\usage{
prewash(expDat, stDat, countDepth = 1000, minGenes = 10,
  alpha1 = 0.01, mu = 1, thresh = 0)
}
\arguments{
\item{expDat}{expDat}

\item{stDat}{sample table}

\item{countDepth}{counts to sample per cell}

\item{minGenes}{min genes}

\item{alpha1}{alpha 1}

\item{mu}{mu}

\item{thresh}{detection threshold for gene stats}
}
\value{
list of expDat, geneStats, okgenes
}
\description{
down samples, computes per gene stats, finds ok genes
}
