% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{probTrans}
\alias{probTrans}
\title{transform mclust class res to expression prob estimates}
\usage{
probTrans(mcRes, expDat, pres = TRUE)
}
\arguments{
\item{mcRes}{what is returned by running Mclust}

\item{expDat}{matrix used as input to Mclust}

\item{pres}{bool whehter to report prob of expression highest 2 classes or not only highest class}
}
\value{
matrix of probabilities of expression class for each cell and gene
}
\description{
transform mclust class res to expression prob estimates
}
