% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{pruneAndScan}
\alias{pruneAndScan}
\title{run dbscan, remove bad points, run again, plot 2 plots, return df of dbscan res}
\usage{
pruneAndScan(tsneRes, eps = 1.8, minPts = 10)
}
\arguments{
\item{eps}{dbscan eps}

\item{minPts}{dbscan minPts}

\item{mtsneRes, }{# has columns TSNE.1 and TSNE.1}
}
\value{
ggplot
}
\description{
run dbscan, remove bad points, run again, plot 2 plots, return df of dbscan res
}
