% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{ptGetTop}
\alias{ptGetTop}
\title{makes vector of gene pairs, iterates over this and computes pairDat, sc_testPattern, then, at the end, findBestPairs}
\usage{
ptGetTop(expDat, cell_labels, topX = 50, sliceSize = 5000,
  ncores = detectCores())
}
\arguments{
\item{expDat}{expDat}

\item{cell_labels}{named vector, value is grp, name is cell name}

\item{topX}{50}
}
\value{
vector of gene-pair names
}
\description{
makes vector of gene pairs, iterates over this and computes pairDat, sc_testPattern, then, at the end, findBestPairs
}
