% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{query_transform}
\alias{query_transform}
\title{makes complete gene-to-gene comparison}
\usage{
query_transform(expDat, genePairs)
}
\arguments{
\item{expDat}{expDat}

\item{genePairs}{genePairs}
}
\value{
matrix indicating which gene of a pair is greater
}
\description{
makes complete gene-to-gene comparison
}
