% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{randomize}
\alias{randomize}
\title{randomize data matrix}
\usage{
randomize(expDat, num = 50)
}
\arguments{
\item{expDat}{expDat}

\item{num}{number of profiles to return}
}
\value{
exp matrix random
}
\description{
randomize data matrix
}
