% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{reorderCellsByGrp}
\alias{reorderCellsByGrp}
\title{re-order cells for plotting}
\usage{
reorderCellsByGrp(nvect, newOrder)
}
\arguments{
\item{nvect}{named vector amed vector of cell-> grp}

\item{newOrder}{vector of grp names in new order}
}
\value{
named vector new order
}
\description{
re-order cells for plotting
}
