% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{rf_classPredict}
\alias{rf_classPredict}
\title{classify samples}
\usage{
rf_classPredict(rfObj, expQuery, numRand = 50)
}
\arguments{
\item{rfObj}{result of running sc_makeClassifier}

\item{expQuery}{expQuery}

\item{numRand}{numRand}
}
\value{
classRes matrix
}
\description{
classify samples
}
