% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_assess.R
\name{sc_Accu}
\alias{sc_Accu}
\title{Assess classifiers based on validation data}
\usage{
sc_Accu(ct_scores, stVal, dLevel = "description1",
  dLevelSID = "sample_name", threshs = seq(0, 1, by = 0.005),
  nRand = 50)
}
\arguments{
\item{ct_scores}{matrix of classification scores, rows = classifiers, columns = samples, colnames=sampleids}

\item{stVal}{sample table}

\item{dLevelSID}{column to indicate sample id}

\item{threshs}{seq of pval cutoffs}

\item{classLevels}{column name of stVal to use as ground truth to assess classifiers}
}
\value{
list of data frames with threshold
}
\description{
Assess classifiers based on validation data
}
