% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter.R
\name{sc_classify}
\alias{sc_classify}
\title{classify data}
\usage{
sc_classify(classList, expDat, cttComms)
}
\arguments{
\item{classList}{result of running CN3_makeClassifier}

\item{expDat}{expression data matrix}

\item{cttComms}{list of genes that were used to train each classifer cttComms[[className]]<-c(gene1, gene2, ...)}
}
\value{
classification matrix nrow=length(classList) ncol=ncol(expDat)
}
\description{
run binary classifiers on input expression matrix
}
