% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_assess.R
\name{sc_classThreshold}
\alias{sc_classThreshold}
\title{determine performance of classification at given class threshold}
\usage{
sc_classThreshold(vect, classification, thresh)
}
\arguments{
\item{vect, }{classifcation score for each given cell}

\item{classification}{actual classification}

\item{thresh}{threshold above which to make a call}
}
\value{
accuracy
}
\description{
determine performance of classification at given threshold
}
