% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sc_filterCells}
\alias{sc_filterCells}
\title{find cells that pass criteria}
\usage{
sc_filterCells(sampTab, minVal = 1000, maxValQuant = 0.95)
}
\arguments{
\item{sampTab, }{which must have UMI column}

\item{minVal}{umis must exceed this}

\item{maxValQuant}{quantile to select max threshold}
}
\value{
vector rownames(sampTab) meeting criteria
}
\description{
based purely on umis
}
