% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sc_findEnr}
\alias{sc_findEnr}
\title{finds genes higher in one group vs others}
\usage{
sc_findEnr(expDat, sampTab, dLevel = "group")
}
\arguments{
\item{expDat}{expression matrix}

\item{dLevel}{dLevel}

\item{list}{of cells in eahc groupsampTab result of running cutreeDynamicTree}
}
\value{
list of grps -> names
}
\description{
finds genes higher in one group vs others
}
