% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sc_hmClass}
\alias{sc_hmClass}
\title{heatmap of the classification result}
\usage{
sc_hmClass(
  classMat,
  grps,
  isBig = FALSE,
  maxPerGrp = 100,
  cRow = FALSE,
  cCol = FALSE,
  fontsize_row = 8,
  scale = FALSE,
  label_reorder = TRUE
)
}
\arguments{
\item{classMat}{classMat}

\item{isBig}{is this a big heatmap? TRUE or FALSE}

\item{cluster_cols}{cluster_cols}
}
\value{
nothing
}
\description{
Heatmap of the classification result.
}
\examples{
cn_HmClass(cnRes, isBig=TRUE)

}
