% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{sc_makeClassifier}
\alias{sc_makeClassifier}
\title{make a classifier, with a randomized class too}
\usage{
sc_makeClassifier(expTrain, genes, groups, nRand = 50, ntrees = 2000)
}
\arguments{
\item{expTrain}{training data}

\item{genes}{vector of genes to use as predictors}

\item{groups}{named vector of cells to groups or classes}

\item{nRand}{=50 num of randomized profiles to make}

\item{ntrees}{=2000 number of trees to build}
}
\value{
RF
}
\description{
rmake a classifier, with a randomized class too
}
