% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_train.R
\name{sc_trans_rnaseq}
\alias{sc_trans_rnaseq}
\title{weighted subtraction from mapped reades and log applied to all}
\usage{
sc_trans_rnaseq(expRaw, total, genes)
}
\arguments{
\item{expRaw}{matrix of total mapped reads per gene/transcript}

\item{total}{numeric post transformation sum of read counts}

\item{genes}{subset of genes to perform this on}
}
\value{
vector of downsampled read mapped to genes/transcripts
}
\description{
down samples
}
