% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnstatus.R
\name{scn_extract_SN_DF}
\alias{scn_extract_SN_DF}
\title{returns a DF of: sample_id, description, ctt, subnet_name, score}
\usage{
scn_extract_SN_DF(scores, sampTab, dLevel, rnames = NULL, sidCol = "sample_id")
}
\arguments{
\item{scores}{a matrix of subNet scores}

\item{sampTab}{sample table}

\item{dLevel}{column name of sampTab to group values by}

\item{rnames}{rownames to extract}

\item{sidCol}{sample identifier column name}
}
\value{
returns a DF of: sample_id, description, ctt, subnet_name, score
}
\description{
returns a DF of: sample_id, description, ctt, subnet_name, score
}
