% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scn_train.R
\name{scn_train}
\alias{scn_train}
\title{Training}
\usage{
scn_train(stTrain, expTrain, dLevel, colName_samp = "row.names",
  nTopGenes = 10, nTopGenePairs = 25, nRand = 70, nTrees = 1000,
  stratify = FALSE, weightedDown_total = 1500,
  weightedDown_dThresh = 0.25, transprop_xFact = 10000)
}
\arguments{
\item{stTrain}{a dataframe that matches the samples with category}

\item{expTrain}{the expression matrix}

\item{dLevel}{the name of the column that contains categories}

\item{colName_samp}{the name of the column that contains sample names}

\item{nTopGenes}{the number of classification genes per category}

\item{nTopGenePairs}{the number of top gene pairs per category}

\item{nRand}{number of random profiles generate for training}

\item{nTrees}{number of trees for random forest classifier}

\item{weightedDown_dThresh}{the threshold at which anything lower than that is 0 for weighted_down function}

\item{transprop_xFact}{scaling factor for transprop}

\item{weightDown_total}{numeric post transformation sum of read counts for weighted_down function}
}
\value{
a list containing normalized expression data, classification gene list, cnPRoc
}
\description{
Tranining broad class classifier
}
