% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnstatus.R
\name{scn_trainNorm}
\alias{scn_trainNorm}
\title{Figure out normalization factors for GRNs, and norm training data}
\usage{
scn_trainNorm(expTrain, stTrain, subNets, classList = NULL,
  dLevel = "description1", tVals = NULL, classWeight = FALSE,
  exprWeight = FALSE, sidCol = "sample_id", xmax = 1000,
  predSD = FALSE)
}
\arguments{
\item{expTrain}{expression matrix}

\item{stTrain}{sample table}

\item{subNets}{named list of genes, one list per CTT, tct=>gene vector}

\item{classList}{list of classifiers}

\item{dLevel}{column name to group on}

\item{tVals}{seful when debugging}

\item{classWeight}{weight GRN status by importance of gene to classifier}

\item{exprWeight}{weight GRN status by expression level of gene?}

\item{sidCol}{sample id colname}
}
\value{
list of trainingScores, normVals, raw_scores, minVals, tVals=tVals
}
\description{
Exactly that.
}
