% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{skylineClass}
\alias{skylineClass}
\title{Skyline waterfall}
\usage{
skylineClass(classRes, cellType, sampTab, dLevel, yval, reorder = TRUE)
}
\arguments{
\item{yval}{}
}
\value{
nothing
}
\description{
Skyline waterfall of classification result
}
\examples{
skylineClass(classRes, "blood",sampTab, "timepoint", 0.25)

}
