% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tsneMultsimp}
\alias{tsneMultsimp}
\title{plot tsne and genes}
\usage{
tsneMultsimp(tsneDat, expDat, genesToPlot, colorPal = "BuPu", revCol = TRUE,
  toScale = TRUE, limits = c(-3, 3))
}
\arguments{
\item{tsneDat}{tsne matrix}

\item{expDat}{expression matrix}

\item{genesToPlot}{genes to plot}

\item{colorPal}{"BuPu"}

\item{revCol}{TRUE}

\item{toScale}{scale?}
}
\value{
ggplot
}
\description{
plot tsne and genes
}
