% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_loadObject}
\alias{utils_loadObject}
\title{loads an R object when you don't know the name}
\usage{
utils_loadObject(fname)
}
\arguments{
\item{fname}{file}
}
\value{
variable
}
\description{
loads an R object when you don't know the name
}
