% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_myDist}
\alias{utils_myDist}
\title{1-PCC distance}
\usage{
utils_myDist(x)
}
\arguments{
\item{x}{numeric matrix}
}
\value{
distance matrix
}
\description{
1-PCC distance
}
\examples{
xdist<-utils_myDist(t(expDat))
plot(hclust(xdist, 'ave'), hang=-1)

}
