% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_strip_fname}
\alias{utils_strip_fname}
\title{reduces full path to filename}
\usage{
utils_strip_fname(str)
}
\arguments{
\item{string}{}
}
\value{
something
}
\description{
reduces full path to filename
}
