% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_stripwhite}
\alias{utils_stripwhite}
\title{strip whitespace from a string}
\usage{
utils_stripwhite(string)
}
\arguments{
\item{string}{string}
}
\value{
new string
}
\description{
strip whitespace from a string
}
