% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chop.R
\name{vg_pca}
\alias{vg_pca}
\title{find var genes and pca}
\usage{
vg_pca(expDat, geneStats, zThresh = 2, meanType = "overall_mean")
}
\arguments{
\item{expDat}{expDat}

\item{geneStats, }{from running sc_statTab}

\item{zThresh}{thresh for defining genes as variable}
}
\value{
list of varGenes and pcaRes
}
\description{
find var genes and pca
}
