% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wash.R
\name{wash}
\alias{wash}
\title{subsets on ok genes, applies selected data transform}
\usage{
wash(preWashed, transMethod = "prop", threshold = 1, removeBad = TRUE)
}
\arguments{
\item{preWashed}{list of expDat=expDat, geneStats=geneStats, okGenes=okGenes from prewash()}

\item{transMethod}{transMethod}

\item{threshold}{threshold if method==binary then values less than threshold get set to 0, everything else gets set to 1}

\item{removeBad}{}
}
\value{
list of geneStats, expDat, transMethod
}
\description{
down samples, computes per gene stats, finds ok genes
}
