% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractSeurat}
\alias{extractSeurat}
\title{extract sampTab and expDat seurat object into regular S3 objects}
\usage{
extractSeurat(seurat_object, exp_slot_name = "counts")
}
\arguments{
\item{seurat_object}{seurat objet}

\item{exp_slot_name}{expression data type}
}
\value{
list
}
\description{
extract sampTab and expDat seurat object into regular S3 objects
}
