% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{findClassyGenes}
\alias{findClassyGenes}
\title{find candidate classifier-worthy genes}
\usage{
findClassyGenes(expDat, sampTab, dLevel, topX = 25, dThresh = 0,
  alpha1 = 0.05, alpha2 = 0.001, mu = 2)
}
\arguments{
\item{expDat}{expDat}

\item{sampTab}{sampTab}

\item{dLevel}{dLevel}

\item{topX}{topX}

\item{dThresh}{dThresh}

\item{alpha1}{alpha1}

\item{alpha2}{alpha2}

\item{mu}{mu}
}
\value{
list of cgenes and grps
}
\description{
find candidate classifier-worthy genes
}
