% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/butter.R
\name{getClassGenes}
\alias{getClassGenes}
\title{getClassGenes}
\usage{
getClassGenes(diffRes, topX = 25, bottom = TRUE)
}
\arguments{
\item{diffRes}{a df with cval, holm, rownames=genes}

\item{topX}{number of genes to select}

\item{bottom}{boolean if ture use the top x genes with - cvals}
}
\value{
vector of genes
}
\description{
extract genes for training classifier
}
