% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pca_to_tsne}
\alias{pca_to_tsne}
\title{make tsne from pca}
\usage{
pca_to_tsne(recRes, perplexity = 30, theta = 0.3)
}
\arguments{
\item{recRes}{result of running gpaRecurse}

\item{perplexity}{(30)}

\item{theta}{(0.30)}
}
\value{
tsne matrix
}
\description{
make tsne from pca
}
