% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsp_rf.R
\name{ptGetTop}
\alias{ptGetTop}
\title{Find the best gene pairs for training}
\usage{
ptGetTop(
  expDat,
  cell_labels,
  cgenes_list = NA,
  topX = 50,
  sliceSize = 5000,
  quickPairs = TRUE
)
}
\arguments{
\item{expDat}{expDat}

\item{cell_labels}{named vector, value is grp, name is cell name}

\item{cgenes_list}{the list of labelled cgenes}

\item{topX}{number of genepairs for training}

\item{sliceSize}{the size of the slice for pair transform. Default at 5e3}

\item{quickPairs}{TRUE if wanting to select the gene pairs in a quick fashion}
}
\value{
vector of top gene-pair names
}
\description{
Find the gene pairs that most distinguish a cancer group from the rest
}
