% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wash.R
\name{weighted_down}
\alias{weighted_down}
\title{weighted subtraction from mapped reades, applied to all}
\usage{
weighted_down(expRaw, total)
}
\arguments{
\item{expRaw}{matrix of total mapped reads per gene/transcript}

\item{total}{numeric post transformation sum of read counts}
}
\value{
vector of downsampled read mapped to genes/transcripts
}
\description{
Simulate expression profile of  _total_ mapped reads
}
